2-SQL/requirements.txt
Django==3.2.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==2.2.2